refactor: move mdx out
