+++ app/design/layout.tsx
@@ -5,7 +5,7 @@ import DesignPageNav from './(components)/page-nav';
 import 'katex/dist/katex.min.css';
 
 import ContainerLayout from '@/components/layouts/container';
-import { Article } from '@/components/templates/mdx';
+import { Article } from '@/components/mdx';
 
 // -----------------------------------------------------------------------------
 // Metadata
