+++ components/mdx/hr.tsx
