+++ components/mdx/math-display.tsx
